requirements.txt
requests
Pillow
schedule